refactor: applicantCount long 타입으로 수정

+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/service/RunnerPostService.java
@@ -235,7 +235,7 @@ public class RunnerPostService {
         return runnerPostRepository.findBySupporterIdAndReviewStatus(pageable, supporterId, reviewStatus);
     }
 
-    public List<Integer> readCountsByRunnerPostIds(final List<Long> runnerPostIds) {
+    public List<Long> readCountsByRunnerPostIds(final List<Long> runnerPostIds) {
         return supporterRunnerPostRepository.countByRunnerPostIdIn(runnerPostIds);
     }
